requirements.txt
opencv-python
jupyterlab
click
matplotlib
pillow
ruamel.yaml